feat: hierarchy

+++ application/controllers/Instrumen.php
@@ -7,11 +7,22 @@ class Instrumen extends CI_Controller {
     public function __construct()
     {
         parent::__construct();
+        cek_login();
         $this->load->model('Instrumen_model');
     }
 
     public function index()
     {
+        $level = $this->session->userdata('level');
+
+        if ($level == 'mahasiswa') {
+            $this->load->view('templates/header');
+            $this->load->view('templates/sidebar');
+            $this->load->view('no_access');
+            $this->load->view('templates/footer');
+            return;
+        }
+
         $data['instrumen'] = $this->Instrumen_model->get_all();
         $this->load->view('templates/header');
         $this->load->view('templates/sidebar');
@@ -19,13 +30,18 @@ class Instrumen extends CI_Controller {
         $this->load->view('templates/footer');
     }
 
-    public function create()
-    {
-        $this->load->view('instrumen/create');
-    }
-
     public function store()
     {
+        $level = $this->session->userdata('level');
+
+        if ($level == 'mahasiswa') {
+            $this->load->view('templates/header');
+            $this->load->view('templates/sidebar');
+            $this->load->view('no_access');
+            $this->load->view('templates/footer');
+            return;
+        }
+
         $data = [
             'namainstrumen' => $this->input->post('namainstrumen'),
             'tahun'         => $this->input->post('tahun')
@@ -37,12 +53,32 @@ class Instrumen extends CI_Controller {
 
     public function edit($id)
     {
+        $level = $this->session->userdata('level');
+
+        if ($level == 'mahasiswa') {
+            $this->load->view('templates/header');
+            $this->load->view('templates/sidebar');
+            $this->load->view('no_access');
+            $this->load->view('templates/footer');
+            return;
+        }
+
         $data = $this->Instrumen_model->get_by_id($id);
         echo json_encode($data);
     }
 
     public function update($id)
     {
+        $level = $this->session->userdata('level');
+
+        if ($level == 'mahasiswa') {
+            $this->load->view('templates/header');
+            $this->load->view('templates/sidebar');
+            $this->load->view('no_access');
+            $this->load->view('templates/footer');
+            return;
+        }
+
         $data = [
             'namainstrumen' => $this->input->post('namainstrumen'),
             'tahun'         => $this->input->post('tahun')
@@ -54,6 +90,16 @@ class Instrumen extends CI_Controller {
 
     public function delete($id)
     {
+        $level = $this->session->userdata('level');
+
+        if ($level == 'mahasiswa') {
+            $this->load->view('templates/header');
+            $this->load->view('templates/sidebar');
+            $this->load->view('no_access');
+            $this->load->view('templates/footer');
+            return;
+        }
+        
         $this->Instrumen_model->delete($id);
         redirect('instrumen');
     }
